Fix the update method in the TorrentService, minor refactoring

+++ src/TransmissionManager.Api/Transmission/Models/TorrentGetResponse.cs
@@ -16,7 +16,7 @@ public sealed class TorrentGetResponseArguments
 
 public sealed class TorrentGetResponseItem
 {
-    public int? Id { get; init; }
+    public long? Id { get; init; }
 
     public string? Name { get; init; }
